prism completed

+++ Assets/Scripts/FiguresOptions/PrismOptions.cs
@@ -5,7 +5,7 @@ namespace Assets.Scripts.FiguresOptions
     class PrismOptions : IOption
     {
         public float height; 
-        public float faces; 
+        public int faces; 
         public float radius; 
 
         public override string ToString() => $"h: {height}, f: {faces}, r: {radius}";
